Moves SecondaryIndex code into `secondary` module
